[domain] separate out error module; reexport either monad
